Model changes
 - Buffered models directly consume IModels
 - Document IModel more
 - Move contraption world render spoofing to flywheel
 - Miscellaneous new RenderMaths
 - Added WorldModel, renders many blocks given a world instance
 - Fix broken transparency on contraptions when using Flywheel

+++ src/main/java/com/jozufozu/flywheel/backend/model/ModelRenderer.java
@@ -1,17 +1,26 @@
 package com.jozufozu.flywheel.backend.model;
 
+import java.util.function.Supplier;
+
+import com.jozufozu.flywheel.core.model.IModel;
+
 public class ModelRenderer {
 
-	protected BufferedModel model;
+	protected Supplier<IModel> modelSupplier;
+	protected IBufferedModel model;
 
-	public ModelRenderer(BufferedModel model) {
-		this.model = model;
+	protected boolean initialized;
+
+	public ModelRenderer(Supplier<IModel> modelSupplier) {
+		this.modelSupplier = modelSupplier;
 	}
 
 	/**
 	 * Renders this model, checking first if there is anything to render.
 	 */
 	public void draw() {
+
+		if (!isInitialized()) init();
 		if (!model.valid()) return;
 
 		model.setupState();
@@ -19,7 +28,21 @@ public class ModelRenderer {
 		model.clearState();
 	}
 
+	protected void init() {
+		initialized = true;
+		IModel model = modelSupplier.get();
+
+		if (model.vertexCount() <= 0) return;
+
+		this.model = new IndexedModel(model);
+	}
+
+	public boolean isInitialized() {
+		return initialized;
+	}
+
 	public void delete() {
-		model.delete();
+		if (model != null)
+			model.delete();
 	}
 }
